requirements.txt
discord.py==2.2.2
pip-chill==1.0.1
python-dotenv==1.0.0
